requirements.txt
paho-mqtt
stringcase